+++ src/routes/Router/Router.tsx
@@ -19,7 +19,7 @@ import { SystemUserOverviewPage } from '@/features/amUI/systemUser/SystemUsersOv
 import { SystemUserDetailsPage } from '@/features/amUI/systemUser/SystemUserDetailPage/SystemUserDetailsPage';
 import { CreateSystemUserPage } from '@/features/amUI/systemUser/CreateSystemUserPage/CreateSystemUserPage';
 import { ReporteesPage } from '@/features/amUI/reportees/ReporteesPage';
-import { ReporteeRightsPage } from '@/features/amUI/reporteeRightsPage_x/ReporteeRightsPage';
+import { ReporteeRightsPage } from '@/features/amUI/reporteeRightsPage/ReporteeRightsPage';
 
 import {
   GeneralPath,
